Changed a lot!
Added the framework to make swappable strategies. Right now it just does nothing new cuz I was not able to figure out the dictionary/word solver. I left that method there but without implementation. Lets talk in the morning to figure it out and if there isnt enough time well bullshit something and get this ready for next time. 

I basically need the methods at the bottom of DictionaryHandler to be filled in. Let me know what you think